maritime-tg-bot/requirements.txt
python-telegram-bot
psycopg2-binary
schedule
requests
python-dotenv